v1.0.0 — General-Availability release: full bug-fix sweep
Rust backend (18 hard compile errors fixed):
- tauri.conf.json: move bundle.plugins.updater to top-level plugins.updater
  (Tauri 2.6+ rejects unknown bundle fields; was breaking generate_context!)
- commands.rs: import SharedState; switch sandbox/telemetry handlers to
  std::result::Result<X, String> (workspace Result<T> is single-generic)
- security/sandbox.rs: add missing 'dirs' crate dep for dirs::home_dir()
- openai_compat.rs: maybe_api_key now returns None when requires_api_key
  is false (was returning the key in both if/else branches — credential
  leak to local providers)
- voice/tts.rs: collapse three identical Linux/Windows/default branches
- ai/tools.rs::apply_diff_minimal: actually apply hunks (was a stub that
  only counted them); use real unified-diff parser
- ai/tools.rs::run_code_eval: enforce timeout_seconds via run_with_timeout
  (was computed then ignored — runaway scripts could hang the agent)
- ai/tools.rs::run_gui_action: fix json! format string interpolation
  (was emitting literal '{e}' instead of the error message)
- memory/entities.rs: use (?i:prefix) so capture group stays case-
  sensitive ('my name is Louis and...' now captures 'Louis' not 'Louis and')
- memory/embeddings.rs: add word-unigram features for multi-word inputs
  so semantic queries ('what is my dog's name?') cross the 0.30 threshold
- calendar/caldav.rs::unfold_lines: keep separator space at fold point
  (was corrupting values with spaces at fold point per RFC 5545 §3.1)
- calendar/caldav.rs::ymd_hms_to_unix: idiomatic iterator over mdays
- ai/agent.rs: use .clamp() instead of .min().max()
- ai/providers/bedrock.rs: drop unused 'client' field (SigV4 is stubbed)
- security/network.rs: gate parse_addr with #[cfg(windows)] (matches call site)
- commands.rs::CsvWriter: replace inherent to_string with Display impl
- computer/safety.rs::expand_tilde: use strip_prefix instead of [2..]
- lib.rs: allow clippy::new_ret_no_self / too_many_arguments / ptr_arg at
  crate level (intentional factory pattern + Tauri ABI)
- 30+ unused imports / dead variables cleaned up across all modules

Tests (was 4 failing, now 104/104 pass):
- calendar::caldav::tests::unfold_continuation_lines
- memory::entities::tests::extracts_name_heuristic
- memory::rag::tests::rag_injects_relevant_facts
- memory::rag::tests::rag_inserts_system_if_missing

GitHub Actions:
- build-release.yml & release.yml: fix target path (workspace Cargo.toml
  is at repo root so artifacts end up in target/<target>/release/bundle/,
  not src-tauri/target/...); walk both paths defensively
- Both workflows: remove continue-on-error from cargo fmt / clippy steps
  (they are now hard gates that pass)
- New lint-and-test job runs fmt + clippy + cargo test --lib before any
  binary is built
- Cache keys fixed so warm caches actually hit

Frontend / docs:
- Sidebar.tsx: stale fallback version '0.8.0' -> '1.0.0'
- Version bumped to 1.0.0 in Cargo.toml, package.json,
  src-tauri/tauri.conf.json, README badge
- CHANGELOG.md: full v1.0.0 entry with the complete fix list
- ROADMAP.md: v1.0 GA section added; next milestone is v1.1

Verification (Rust 1.97.1):
- cargo check ............................ passes (0 errors)
- cargo clippy --all-targets -- -D warnings  passes (0 warnings)
- cargo test --lib ....................... 104/104 pass
- cargo fmt --all -- --check ............. passes
- npm run build .......................... passes (296 kB JS bundle)
- npx tsc --noEmit ....................... passes (0 TS errors)

+++ src-tauri/src/memory/knowledge.rs
@@ -67,7 +67,7 @@ impl KnowledgeBase {
             // Bump use count.
             conn.execute(
                 "UPDATE knowledge SET use_count = use_count + 1, last_used_ms = ?1 WHERE key = ?2",
-                params![time::OffsetDateTime::now_utc().unix_timestamp() as i64 * 1000, key],
+                params![time::OffsetDateTime::now_utc().unix_timestamp() * 1000, key],
             )?;
             return Ok(Some(KnowledgeEntry {
                 key: row.get(0)?,
@@ -124,17 +124,15 @@ impl KnowledgeBase {
             "SELECT key, value, source, confidence, created_at_ms, last_used_ms, use_count FROM knowledge",
         )?;
         let rows = stmt.query_map([], |row| {
-            Ok((
-                KnowledgeEntry {
-                    key: row.get(0)?,
-                    value: row.get(1)?,
-                    source: row.get(2)?,
-                    confidence: row.get(3)?,
-                    created_at_ms: row.get::<_, i64>(4)? as u64,
-                    last_used_ms: row.get::<_, i64>(5)? as u64,
-                    use_count: row.get::<_, i64>(6)? as u64,
-                },
-            ))
+            Ok((KnowledgeEntry {
+                key: row.get(0)?,
+                value: row.get(1)?,
+                source: row.get(2)?,
+                confidence: row.get(3)?,
+                created_at_ms: row.get::<_, i64>(4)? as u64,
+                last_used_ms: row.get::<_, i64>(5)? as u64,
+                use_count: row.get::<_, i64>(6)? as u64,
+            },))
         })?;
 
         let q_tokens: std::collections::HashSet<String> = query
